Handle submission merge failures caused by outdated fork
The automatically generated access token provided by `${{ secrets.GITHUB_TOKEN }}` is used to automatically merge
submission pull requests if they are compliant with all requirements.

If the pull request's fork is behind the parent repository and any GitHub Actions workflow has been modified since that
time, the token permissions are downgraded, which causes the GitHub API request for merging the PR to fail with a 403
status. Previously, this was treated as an unexpected merge failure caused by some problem not resolvable by the PR
author. Since this issue can easily be resolved by the PR author by bringing their fork up to date, the "Manage PRs"
workflow is hereby changed to provide instructions for doing so. As before, a review will be requested from the
maintainer so that they can monitor the situation and provide the PR author with assistance if needed.